report: Add --platform flag for Hyper-V
Currently, with Hyper-V guests, the report_data of an attestation report
is generated beforehand by the platform. This indicates that a user
cannot write some optional data to be included in the report. With
--platform enabled, the pre-generated platform data will be written to
the file pointed to by $REQUEST_PATH (or "platform-request-file.txt" if
no $REQUEST_PATH is given as input).

Signed-off-by: Tyler Fanelli <tfanelli@redhat.com>