VLEK: Adding ASVK Support
With the release of the VLEK, there is a separate
api endpoint for requesting VLEK signing keys which
are known as an AMD Signing VLEK Key (ASVK). This
adds logic to handle these special use-cases. As
the GHCB does not offer a unique UUID, assumptive
logic must be applied to when the file should be
named appropriately ASVK or ASK.

Signed-off-by: Larry Dewey <larry.dewey@amd.com>