Updating to version 6 of library and fixing attestation
This update bumps the dependency to version 6 of the SEV libarary.

Besides fixing the code compatibility, we changed the way certificate
fetching is done.

Now the user is expected to provide a V3 attestation report, with which
we can tell for what processor model we will be getting the certificates
from. If the user doesn't have a V3 report in the firmware, or wants to
get certs for a different proc model, they can still use the
--processor_model flag.

The verification also add a --processor_model flag in case they need to
attest a V2 of the report.

Signed-off-by: DGonzalezVillal <Diego.GonzalezVillalobos@amd.com>