Updating to version 6 of library and fixing attestation (#89)
This update bumps the dependency to version 6 of the SEV libarary.

Besides fixing the code compatibility, we changed the way certificate
fetching is done.

Now the user can provide a V3 attestation report using the --report flag, that can be used to tell what processor model to use to get the appropriate certificates.
If the user doesn't have a V3 report, or wants to
get certs for a different processor model, they can still pass in the
desired proc model as default behavior.

The verification also add a --processor_model flag in case they need to
attest a V2 of the report.

Signed-off-by: DGonzalezVillal <Diego.GonzalezVillalobos@amd.com>