Removing intird and append requirement for kernel measurements
Had initrd and append as required when using a kernel for measurement.
There are use-cases where the just the kernel is used (uki images), so
we remove the initrd and append as required when using kernel. Kernel is
still needed if user is passing either the initrd or the append.

Signed-off-by: DGonzalezVillal <Diego.GonzalezVillalobos@amd.com>