fix(hyperv): downgrade VMPL check from error to warning
Currently the code enforced VMPL 0 by returning an error when a different
value was specified via --vmpl. However, since the attestation report is
always fetched from the vTPM at VMPL 0 when using the --platform flag,
the --vmpl argument has no actual effect in this context.

This change replaces the error with a warning.

Signed-off-by: Takuma IMAMURA <209989118+hyperfinitism@users.noreply.github.com>